PR Refactoring
[CSHRCP-311]
