+++ location-service/src/main/java/uk/gov/cshr/locationservice/service/GoogleService.java
@@ -15,7 +15,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import uk.gov.cshr.locationservice.LocationServiceException;
-import uk.gov.cshr.locationservice.RegionLookup;
 import uk.gov.cshr.locationservice.controller.Coordinates;
 
 @Service
